chore(MyKSuite): Rename package with new Core's standard
