+++ MykSuite/src/main/java/com/infomaniak/core/myksuite/ui/components/WeightOneSpacer.kt
@@ -1,0 +1,36 @@
+/*
+ * Infomaniak Core - Android
+ * Copyright (C) 2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.core.myksuite.ui.components
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.Dp
+
+// TODO Add in CoreUi
+@Composable
+fun ColumnScope.WeightOneSpacer(minHeight: Dp) {
+    Spacer(Modifier.height(minHeight))
+    Spacer(Modifier.weight(1.0f))
+}
+
+@Composable
+fun RowScope.WeightOneSpacer(minWidth: Dp) {
+    Spacer(Modifier.width(minWidth))
+    Spacer(Modifier.weight(1.0f))
+}
