+++ MykSuite/src/main/java/com/infomaniak/core/myksuite/ui/theme/Margin.kt
@@ -1,0 +1,37 @@
+/*
+ * Infomaniak Core - Android
+ * Copyright (C) 2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.core.myksuite.ui.theme
+
+import androidx.compose.ui.unit.dp
+
+object Margin {
+    /** 4dp */
+    val Micro = 4.dp
+    /** 8dp */
+    val Mini = 8.dp
+    /** 12dp */
+    val Small = 12.dp
+    /** 16dp */
+    val Medium = 16.dp
+    /** 24dp */
+    val Large = 24.dp
+    /** 32dp */
+    val Huge = 32.dp
+    /** 48dp */
+    val Giant = 48.dp
+}
