+++ MykSuite/src/main/java/com/infomaniak/core/myksuite/ui/views/MyKSuiteChipView.kt
@@ -1,0 +1,28 @@
+/*
+ * Infomaniak Core - Android
+ * Copyright (C) 2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.core.myksuite.ui.views
+
+import android.content.Context
+import android.util.AttributeSet
+import com.infomaniak.core.myksuite.ui.components.MyKSuiteTier
+
+class MyKSuiteChipView @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0,
+) : BaseMyKSuiteChipView(context, attrs, defStyleAttr, MyKSuiteTier.Free)
