--- MykSuite/src/main/java/com/infomaniak/lib/myksuite/ui/theme/Dimens.kt
@@ -1,34 +1,0 @@
-/*
- * Infomaniak Core - Android
- * Copyright (C) 2025 Infomaniak Network SA
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.infomaniak.lib.myksuite.ui.theme
-
-import androidx.compose.ui.unit.dp
-
-object Dimens {
-
-    /** 56 dp */
-    val buttonHeight = 56.dp
-    /** 16 dp */
-    val smallIconSize = 16.dp
-    /** 24 dp */
-    val iconSize = 24.dp
-    /** 16 dp */
-    val largeCornerRadius = 16.dp
-    /** 5 dp */
-    val cardElevation = 5.dp
-}
