--- MykSuite/src/main/java/com/infomaniak/lib/myksuite/ui/views/MyKSuiteDashboardFragment.kt
@@ -1,51 +1,0 @@
-/*
- * Infomaniak Core - Android
- * Copyright (C) 2025 Infomaniak Network SA
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.infomaniak.lib.myksuite.ui.views
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import androidx.navigation.fragment.navArgs
-import com.infomaniak.lib.myksuite.ui.screens.MyKSuiteDashboardScreen
-
-open class MyKSuiteDashboardFragment : Fragment() {
-
-    private val navigationArgs: MyKSuiteDashboardFragmentArgs by navArgs()
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        return ComposeView(requireContext()).apply {
-            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-            setContent {
-                val onClose: () -> Unit = { this@MyKSuiteDashboardFragment.findNavController().popBackStack() }
-                with(navigationArgs) {
-                    MyKSuiteDashboardScreen(
-                        userName = userName,
-                        avatarUri = avatarUri,
-                        dailySendingLimit = dailySendLimit,
-                        onClose = onClose,
-                    )
-                }
-            }
-        }
-    }
-}
